requirements.txt
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cupy-cuda12x==12.1.0
cycler==0.12.1
datasets==2.18.0
dill==0.3.8
diskcache==5.6.3
docker-pycreds==0.4.0
docstring_parser==0.16
eval-type-backport==0.1.3
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fastapi==0.110.0
fastrlock==0.8.2
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.21.4
idna==3.6
importlib_resources==6.3.0
interegular==0.3.3
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lark==1.1.9
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opacus==1.4.1
opt-einsum==3.3.0
outlines==0.0.36
packaging==24.0
pandas==2.2.1
peft==0.10.0
pillow==10.2.0
prometheus_client==0.20.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.1
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rank-bm25==0.2.2
ray==2.9.3
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
sentencepiece==0.2.0
sentry-sdk==1.45.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.1.2
tqdm==4.66.2
transformers==4.38.2
triton==2.1.0
trl==0.8.5
typing_extensions==4.10.0
tyro==0.8.3
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
vllm==0.3.3
wandb==0.16.6
watchfiles==0.21.0
websockets==12.0
xformers==0.0.23.post1
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.0
